refactor(coordinator): 抽出 llm_pipeline 子模块
机械式行为保持拆分：从 coordinator.rs 搬出 LLM 润色/翻译/provider 构建/凭证读取管线
—— 流式与一次性润色、润色+翻译合并、各 ASR/LLM provider 凭证读取与 ActiveLLMProvider
构建、provider 分类器、中文字形偏好、词库 phrases/hotwords 读取。polish-translate
marker 常量经 glob 重导出供父模块 tests 的 super:: 引用。
